SelectableLabel: Use onPressedChanged
OnTapped does not immediately show the context menu when label is
rightclicked, breaking QQC2 CI in the process. OnTapped shows the context
menu *after* right click is released.

This changes it to be similar to other elements we have like TextField
where right click brings the context menu immediately, instead of on
right click button release.

We utilize onPressedChanged, then check if pressed is true, to pop up the
context menu immediately.

CCBUG:481293