Fix mobile mode
In Qt6 clicks are always stolen from the background swipe items,
breaking the toolbuttons. This is fixable by putting them in front
on the main contentitem, drawing them in a way that still looks
like they're behind.
this also makes it possible to use transparent backgrounds for the list item

No visual changes are made