Always use a ToolBar for pages on the stack
On mobile, usage of the stack is quite common. For example, opening the
settings in Tokodon. However, due to this check - page actions disappear
because it's in Title mode.

Instead, we should always use the ToolBar style when determining the
global ToolBarStyle for a stack. This makes page actions show up on
mobile when pushing pages via the layers system.