PageRow: Fix the navigation buttons disappearing in certain situations
In Tokodon we have a page that has the header explicitly disabled, and
then we push onto it a layer that has it's header enabled. This works
as expected, except that the navigation buttons disappear which causes
some issues as you can never go back.

Looking further, it seems to be due to some weird and convoluted active
checks in globalToolBarUI. We only need to make sure it's active if the
header isn't disabled completely, *or* the leading visible item has it
enabled explicitly.