Mnemonic: Emit sequence change also when it gets disabled
When merely changing enabled to false, the seuqence is reset but the
property is not notified. This results in the Shortcut item on the
control to still have the shortcut and might cause ambiguous shortcuts.

For example, in case of Kirigami ActionToolbar where there's usually
two instances (icon and full) that are shown/hidden as needed, both
buttons will remain active as disabling the mnemonic didn't work.