FormLayout: Set mnemonic only if buddy is visible and doesn't have one
Reduces the number of dead or redundant mnemonics.

For example, with a CheckBox as buddy, the CheckBox will already have
an mnemonic on its own, no need to have one on the form label next to it,
too. Also, for example when the control is replaced by a static label
conditionally (e.g. the "single refresh rate" in KScreen), the mnemonic
makes no sense if the buddy control is not visible.