controls/Card: set title as default Accessible.name
The card element has a banner that contains a title element, which is
intended as the visible title for that card. This is adequate for
sighted users, but is not passed to the a11y interface, so screen
readers are unaware of its function and default to not reading it out.
Currently, applications using Card have to set this link manually, and
usually do not do so, leaving these controls inaccessible. Given the
typical visual prominence of the banner title for sighted users, this
seems like a sensible default assignment. In the rare cases where this
is less appropriate, applications can still overwrite this manually.
Given the variety of functions that Cards can have, Accessible.role
probably still should be set manually and explicitly.

This change sets the default Accesible.name to the banner title.

CCBUG: 497971