Add an on-by-default "external link" icon for UrlButton
Part of https://invent.kde.org/plasma/discover/-/issues/27

Most URL links will open in an external browser, making them "external
links". It's helpful to show the standard "arrow coming out of a box"
icon for these, both to help users familiar with what that icon means,
and also to accustom those who aren't familiar with it yet.

This icon can be disabled if it's not applicable (e.g. if it's used in
an actual web browser where the link opens inline, not externally).