SwipeListItem: improve keyboard navigation
SwipeListItem is commonly used as a delegate for ListViews that have
actions associated with them. On desktop, their keyboard navigability is
somewhat awkward - to reach them with the tab key, the user has to
traverse the tab chain until reaching that button, irrespective of the
currently selected list entry. Arrow keys do not work either.

This change lets the user skip immediately to the currently selected
entry's actions when pressing the tab key. Pressing the arrow key in the
direction of the action button also focuses that button, and arrow keys
on the buttons themselves work as well (to select different buttons, or
move up/down the list).