Dialog: include leading and trailing components in footer implicitWidth
4fb4d656dfa5161ac4676d278bcd7cb08327a753 fixed footer sizing by taking
into account the implicitWidth of the DialogButtonBox, which can be
wider that the main body content. However leading and trailing footer
components were missed, so if any of those are included in a particular
Dialog's implementation, its footer may still overflow.

Account for those by getting the implicitWidth of the RowLayout item
that holds all of them, which also fixes the spacing between items not
being included in the calculation.

BUG: 497795
FIXED-IN: 6.11