ScrollablePage: add ensureVisibility method
If the ScrollablePage's contents do not automatically handle positioning
the viewable area on focus changes (the way that Views generally do),
passing focus with e.g. keyboard navigation may leave the currently
focused item out of view. Applications filling the page directly with
e.g. a FormLayout or CardsLayout will need to scroll explicitly when
necessary. This is possible with ScrollablePage's API, but the scrolling
logic needs to be implemented by each application separately.

This change adds a convenience function that scrolls the viewable area
in a reasonable manner, which controls can call if they receive focus.