ActionToolBar: make overflow button ID as ButtonMenu
When the ActionToolBar needs to show an overflow button, the
automatically set a11y role for this is that of a standard button. I
opens a menu though, so the semantically more appropriate ButtonMenu
role  should be used to make the interface more transparent for screen
reader users.

This change sets the Accessible.role of the button to
Accessible.ButtonMenu.