ToolBarPageHeader: pass focus to first button
Page exports the toolbar as an item, but not the individual action
buttons. Consumers that want to put focus on the toolBar may try to
place activeFocus there; for example, ScrollViewKCM from KCMUtils does
this on an up-arrow key press that's not consumed otherwise, such as at
the top of a ListView. It doesn't actually work currently, though,
ToolBarPageHeader itself is not interactable and there is no highlight,
so keyboard focus seems to vanish into thin air. There also does not
seem to be a clear way to set focus on actual tool bar buttons from the
outside (except through accelerator keys).

This change makes it so that if the ToolBarPageHeader receives active
focus, it passes it to the first visible child, which is the first
visible action in reading direction.