TabletModeWatcher: Delay the first call to the portal
TabletModeWatcher construction can be triggered early in the
startup of an app when using Breeze as it checks tablet mode
in a pixel metric. The new portal host registry only allows
registering before any other portal call so delay until the
app is running to give a chance for registering.