LinkButton: give it a pressed effect
Pretty much all of our other controls have a press effect, and real web
links in browsers do too. Let's add the standard-style effect here.

This is implemented by creating three new public properties that clients
are expected to override if they want to change the color; changing the
`color` property directly will break the logic and is not recommended.