SwipeListItem: handle not being direct delegate
The keyboard navigation improvements in b35dbd09 assume that the
SwipeListItem is a delegate, as they need to set the currentIndex in
response to arrow key interaction (i.e. pressing up from a SLI's buttons
should make the previous list item the currentitem, even if the SLI
is not the currentItem (e.g. from tabbing though the controls). This
fails when the SLI is a child of the delegate, rather than the delegate
itself, as the ListView attached property is not available.

This change makes the SwipeListItem search through the parent chain to
find the ListView, and handles the case where it's used outside a
ListView more gracefully.