mnemonicattached: add plainTextLabel property
mnemonicattached provides access to the label in various forms: rich
text, with & markup and original as supplied. There isn't a convenient
way to access the label as shown on the button while inactive though.
This would be helpful in cases where no extra markup is desired, such
as for setting Accessible.names.

This change adds this property.