controls, layouts: set Accessible.name for Mnemonics
Kirigami automatically sets mnemonics for many types of
controls and buddy FormData.labels, which interactively replace the text
with a rich text version that underlines the accelerator key.
The Accessible.names of these labels, however, are implicitly tied to
the text, and so they are updated to the version with rich text
annotation as well. This is without immediate bad consequences, as it is
hard to get orca in a situation where it will read out the label while
the user presses alt. But it is pointless busywork for orca, unnecessary
dbus chatter, makes debugging a11y issues with event logs more annoying,
is not what QtWidgets or Gtk apps do, and if a user somehow manages to
achieve a readout, they will hear a version with  markup sprinkled
in.

This change explicitly sets the Accessible.name of these items with
MnemonicData attached to their plain text representation. Along the way,
we make sure that the controls in Kirigami that use MnemonicData
set the plainTextLabel as their own Accessible.name, in case the user
supplies predefined mnemonics using "&".