layout/FormLayout: scroll to focus item if necesary
QtQuick does not automatically scroll to items that gain focus from e.g.
tab navigation or arrow key navigation. This makes keyboard-based use
hard, as the items being interacted with may be off-screen. A method was
added to ScrollablePage that can handle this, but consumers need to
explicitly call it. In the case of FormLayout, however, we can handle
this centrally, which should take care of a large number of cases where
this would be relevant.

This change introduces a handler that listens to changes in the window's
activeFocusItem, and scrolls to that item if it is a child of the
FormLayout.