InlineMessage: fix link opening and hover handlers
The Label and ActionToolbar occupy the same space, with the
ActionToolbar being rendered last. This makes its own hover handling
override that of the Label, causing the pointing finger cursor to not
appear when hovering a link.

There's nothing wrong with both occupying the same space, but we do need
the Label to be rendered on top so its hover handling takes precedence.

BUG: 500578
FIXED-IN: 6.12