Dialog: fix width and height binding loops
We're doing a complicated calculation to get the implicit widths and
heights of child items, but this is unnecessary because Qt provides
properties we can use for the same purposes. Let's use them.

Tested every Dialog implementation in System Settings, Info Center,
System Monitor, Discover, NeoChat, and Elisa. I found no regressions
with sizing or visual display in any of them, and the binding loop
warning is now gone.

BUG: 494420
FIXED-IN: 6.12

cc @mart @carlschwan @nicolasfella