Dialog: make header contents composable
This commit introduces two components that a Dialog's header is built
out of. To have a dialog with a fancy custom header, you can override
the header: property and build it out of these pieces, plus your own
content. Consider it an alternative to the approach outlined in
https://invent.kde.org/frameworks/kirigami/-/merge_requests/1730.

This approach improves composability and still allows developers to
build fancy Dialog headers without having to re-invent the wheel.