dialogs/DialogHeaderTopContent: remove margin instead of adding
The close button in DialogHeaderTopContent contains some logic for
distinguishing between short and tall dialog headers, to place the
button
in the top right corner for tall headers only. What it actually does is
add extra padding though, so the button is moved toward the bottom,
rather than the top.

Instead, subtract the margin (and use the value from the header instead
of a fixed largeSpacing), and do it for both top and right margins. This
places the close button right at the edge of the popup, similar to how
it looks in task manager tooltips.