DialogLayer: force header style when on global pagestack
when using pushDialogLayer on mobile the page is positioned into
either a Dialog or as a new layer in the pagerow's layers, depending from
the space.

On desktop it creates an ApplicationWindow which is then used as a top level dialog

ideally the page does have its own header when in the main pagerow stack,
otherwise the back button goes over the contents, but not having the header when inside a Dialog.
In desktop mode when in an external ApplicationWindow i guess the presence of the header
depends from the apps preference