HeaderFooterLayout: introduce spacing property
HeaderFooterLayout lays out a main item with an header and a footer
but it's quite limiting that is not possible to specify
a space between the header and the main item and between the footer
and the main item.

To have that is possible to use a Padding component as contentItem,
but since this is intended to be a lightweight layout and needing
a spacing seems a quite common pattern, i think is better having
one object that does the desired layout rather than having to nest
different items instead.

This introduces a new spacing property which spaces of a given
amount of pixels the header and footer from the content item