controls: Improve calculation of InlineMessage implicit height
If the text isn't long enough to wrap into mulitple lines, but we have
many or large actions such that the actions are still put below the
text, the actions and close button would overlap because the close
button is not accounted for. To fix that, always anchor the actions at
the bottom and calculate a more proper implicit height for the entire
inline message that accounts for close button height if it is visible.