SwipeListItem: fix visibility of actions buttons
SwipeListItem sets the visibility of the actions buttons layout
depending on the visibility of the individual buttons; more precisely,
a handler for the individual buttons' onVisibilityChanged triggers sets
the layout's visibility. There is a problem here though: if no buttons
are visible at one point and the layout is invisible, QQuickItem will
not actually emit visibleChanged anymore, as children of an invisible
parent are always invisible as well. If there is e.g. a set of buttons
that alternate in visibility, correct display is based on the arbitrary
order that bindings resolve: if the newly visible items have their
bindings evaluated first, all is well, but if the newly invisible
buttonsgo first, there will be a moment where none are visible and the
whole thing turns off permanently.

Instead, handle updating the layout directly in the binding for visible,
which is always executed.