Polish layer animations
When pushing a layer the animation looks quite janky,
refactor a bit how things move, to have every frame
in the animation to make sense.

when pushing a layer, the whole page with the toolbar slides in going to cover the previous toolbar.
also pay attention to not move in a janky way


This approach also simplifies code a lot.

there are still some ugly things in the code, but a separate refactor of the global toolbar is on the way for a future mr (probably back buttons are also going to be inside the toolbar as a copy for each page rather than be on a top layer, but can't be done in this mr).