Rework the dialog header component to not make it depend on specific parameters in Kirigami.Dialog
Coming out of the conversation here https://invent.kde.org/frameworks/qqc2-desktop-style/-/merge_requests/455#note_1205603

The `DialogHeaderTopContent.qml` component was made dependent on `Kirigami.Dialog`'s API (specifically `showCloseButton`) which meant it can't be used generically. This reworks it. Including the naming so it can now be used in a `QQC2.Dialog` but also as a generic content component elsewhere.

It could potentially be even more generic, think `TitleButtonContent`