Ensure non-visible bottom toolbars don't take up any space
It's apparently possible for the bottom toolbar loader height to be
non-zero even when not being active. This then results in PageRow
accounting for that in the anchor bottom margin for the page content,
resulting in unnecessary empty space at the bottom.

Testcase: Itinerary's favorite location map page. The extra space is
particularly easy to spot there as the map doesn't expand all the way
to the bottom.