Add a polyfill module with a dummy implementation of Qt 6.9's SafeArea
This allows us to use the SafeArea attached property in QML code without
requiring Qt 6.9 yet. On Qt 6.9 our replacement is no longer provided and
we automatically fall back to the real one from QtQuick.

The fact that this will only return 0 margins is no problem in practice,
as the only platform where we already get proper values so far is Android,
which we control the Qt version for our packages anyway and thus will
always use Qt 6.9. We'll also eventually want this for Linux, but once
we have the necessary infrastructure in place in KWin etc those will likely
require Qt 6.9 as well.