controls/PageRow: don't announce StackView over screen readers
PageRow has an internal StackView; it defaults to the Accessible.role
LayeredPane, which is by default announced by screen readers (or at
least by orca). This may be correct in some circumstances, but not here:
it's a purely internal object that has no user relevance; hearing
LAYERED PANE ZERO ITEMS every time a page in a PageRow is entered is
confusing to screen reader users, they have typically no idea what it
means, and if they do there is little they can do with this
information. (I guess it indicates whether a modal layer is open, but in
practice I couldn't seem to actually get this to work, and it seems
better to indicate when a modal layer is open, not every single time
when there isn't one - and that is a separate issue.)

Instead we can set its role explicitly to Pane, that marks it as a
filler in Accerciser and orca does not read it out every time.