Always update render node rect
The render node geometry is influenced by the bounding rectangle of the
Icon item and the image size. m_sizeChanged only accounts for the bounding
rectangle changes, so we miss updating the render node rect when the
image size changes.

calculateNodeRect() is pretty cheap, so we can just drop updateRect.
It'll fix an issue where icons are stretched and simplify the code a bit.

BUG: 504526