UrlButton: fix icon positioning
This commit fixes the icon positioning calculations in `UrlButton` when
the button width is different from the default width, and when layout
mirroring is enabled. The previous icon positioning calculations made
incorrect assumptions about the button's width, and mirroring did not
work due to `LayoutMirroring.enabled` not being propagated to the `Icon`.