Global toolbar refactor: navigation buttons
Change how global navigation buttons are shown and loaded: load a copy on each page global toolbar, shown and hidden by activating and deactivating a Loader.

Even if instantiates a bit more things, simplifies  the logic a lot and makes the animation a bit more sensible, especially for pages pushed in layers, which now the back button will be shown sliding together the toolbar and we don't have to use anymore weird state machine based animations to have the procedure not completely janky.

Now pressing the back button will have the button sliding away as well together with the page, tough is coherent with Android that does the same thing