OverlayDrawer: Supports Mouse resize
When enabled, sidebars that are done with drawers, can be resized
with the mouse between a given minimum and maximum size.

Resize is supported on drawers from all 4 edges.

For retrocompatibility, the feature is off by rdefault, it can be
enabled with the property interactiveResizeEnabled.

It is resposibility of the application implementation to save and
restore the size.

Supports https://invent.kde.org/plasma/discover/-/merge_requests/1128 and https://invent.kde.org/plasma/plasma-systemmonitor/-/merge_requests/376