PageRow: proper width for sidebar background
The sidebar background when is reparented from an OverlayDrawer
is not correctly resized, it keeps the width the Drawer gave to it.

Instead of reparenting the individual items, reparent the popupItem itself, making everything behave much better.

As a bonus, now drawers that have interactiveResizeEnabled will just work also when they are in this sidebar mode