Don't relayout during item destruction
a reason forgetItem is called is when the item is being destroyed.
in that case is really important layoutITems doesn't get called until
the item is actually removed from the list of known items.

header::visiblechanged is connected to layoutitems to we need
to disconnect it before hiding the header

BUG:507483