primitives: Use pre-multiplied blending in ShadowedRectangle shader
The entire shader assumes we are using pre-multiplied colors, because Qt
generally wants pre-multiplied. For blending the texture on top of the
rectangle color, we also need to account for the source texture to be
premultiplied. Since the `sdf_render()` function uses `mix()`
internally, it cannot do this by itself, instead we first need to render
the rectangle with the texture as color, and then blend that on top of
the background.