platform: Skip update() in PlatformTheme if we don't have a window
Otherwise we risk emitting change signals for things while we're in the
middle of teardown. If we don't have a window we're not going to be
visible anyway so we may as well not do anything.

This also moves the check from `windowChanged()` to `update()` to make
sure we check at the right moment.

BUG: 493921