globaltoolbar/NavigationButtons: fix Back button in RtL
The Back button tests whether it should be enabled using the pageStack's
contentX. In LtR, this works, but in RtL it does not, as the top-most
page is also the left-most and therefore has an x-axis position of 0. As
a consequence, the Back button is always disabled in RtL.

Instead, we make this test conditional on LayoutMirroring, keeping the
current method for LtR, and testing whether the position is at 0 and the
total contentWidth of the page stack is wider than the pageStack (and
therefore some parts of the stack must be outside the visible area).

BUG: 511295
FIXED-IN: 6.21