Card: Remove default value documentation for actions
First, we can use a \default command here instead of typing it out
manually. But this only works for QML properties, not aliases like this.
But really this doesn't need to be documented at all, I think its
obvious that its empty by default.

I also added a full-stop to the brief command too, which removes a QDoc
warning.