OverlayDrawer: Move handles in the toolbar
Put directly handles in the top toolbar in order to make a less glitchy page transition animation, simpler code and better themeability with Union.

The handle is a normal toolbar button in the global toolbar, the floating drawer close button is a separate non-flat toolbar button in the previous place of implemetnation, which should be easier to theme with Union.

It drops the fancy animations for the menu button of the hamburger morphing into an x, but on the upside the proper icons are used now, so it looks more consistent with QWidget applications and that icon will come from the theme so looks less broken on 3rd party ones.

The page switch animations for pages and drawers now will animate the drawer handles as well, making it look more "solid"