controls/AboutItem: adjust heading stylings
The AboutItem displays both name and description of the application on
subsequent rows. The only difference is that the description has its
level set to 2 rather than 1, which is only a small difference and does
not make the auxiliary role of the description apparent.

Instead, set the heading level to 3, and mark it as secondary type. This
improves the clarity of the display, and matches what the Kirigami
Addons version does, and what Applet configuration pages are moving to.