controls/AboutItem: align author buttons in fixed column
The authors section in AboutItem displays the authors email and web
links, as available. If not available, the buttons are set to be not
visible. This means that the equivalent button (typically email, which
is apparently universally available) may appear in a different positions
across authors, which makes the table look noisy.

Instead we keep the buttons visible, but set them to disabled and with
opacity 0 if not available. This would change the button alignment from
right-aligned to left-aligned, and make the overall item width look
internally unaligned in the common case of not having weblinks, so we
also change the order to place the email last.