qualify access to parent properties
qmllint will complain if the property of a parent is accessed without
being qualifie by the parent's id.

This change adds the parent id to (almost) all cases where it is
currently absent, giving the parent an id in the small number of
cases where it did not have one.