make action handling linter-friendly
Many components accept both QQC2 Action and its more powerful subtype,
Kirigami.Action. We thus have to test each time it's relevant whether an
Action is an instanceOf Kirigami.Action, and handle it accordingly. This
works fine, but tooling does not understand this and will still complain
that Kirigami-specific properties don't exist on the base type. It's
also somewhat complicated, as to figure out the default value for QQC2
actions we have to parse the logic; it's not explicitly stated.

Instead, we ask forgiveness rather than permission, boldly cast it to
Kirigami.Action, and handle the QQC2 case as the nullish fallback. This
makes qmllint happy, and is also typically shorter and clearer.