controls/Action: fix alternate shortcut
If an Action imported from a QAction has an alternate shortcut set, it
calls root.triggered. The Action root, however is not called root, so
this will look up the closest root id and attempt to trigger it. But
there's no guarantee that the closest root id even has a trigger method,
or that it will happen to do the right thing.

Instead, just call the action by its actual id: action.