controls/GlobalDrawer: explicitly pass backItem
The delegate uses the backItem (from the header in a multi-level global
drawer) by id for some of its operations. They're in the same file, but
both are in separate components, with one being instantiated from the
other. The lookup can only work at runtime, but qmllint at compile time
fails.

Instead, explicitly pass it as a property to the delegate.